+++ src/CMakeLists.txt
@@ -11,8 +11,8 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-set(SOURCES main.c wss_mgr.c mutex.c networking.c nopoll_helpers.c ParodusInternal.c 
-    string_helpers.c time.c config.c connection.c spin_thread.c client_list.c service_alive.c)
+set(SOURCES main.c mutex.c networking.c nopoll_helpers.c nopoll_handlers.c ParodusInternal.c 
+    string_helpers.c time.c config.c connection.c spin_thread.c client_list.c service_alive.c upstream.c downstream.c)
 
 add_executable(parodus ${SOURCES})
