+++ src/connection.h
@@ -1,59 +1,47 @@
+/**
+ * @file connection.h
+ *
+ * @description This header defines functions required to manage WebSocket client connections.
+ *
+ * Copyright (c) 2015  Comcast
+ */
+ 
 #ifndef _CONNECTION_H_
 #define _CONNECTION_H_
 
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <ctype.h>
-#include <unistd.h>
-#include <cJSON.h>
-#include <nopoll.h>
-#include <sys/sysinfo.h>
-#include "wss_mgr.h"
-#include <pthread.h>
-#include <math.h>
-#include <nopoll.h>
-
-#include <netdb.h>
-#include <arpa/inet.h>
-#include <getopt.h>
-#include <signal.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <wrp-c.h>
-#include <nanomsg/nn.h>
-#include <nanomsg/pipeline.h>
-
-#include "ParodusInternal.h"
-#include "time.h"
-#include "config.h"
-#include "parodus_log.h"
+#include "nopoll.h"
+#include "upstream.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-    
+
+/*----------------------------------------------------------------------------*/
+/*                            File Scoped Variables                           */
+/*----------------------------------------------------------------------------*/
+
+extern UpStreamMsg *UpStreamMsgQ;
+
+/*----------------------------------------------------------------------------*/
+/*                             Function Prototypes                            */
+/*----------------------------------------------------------------------------*/
+
+/**
+ * @brief Interface to create WebSocket client connections.
+ * Loads the WebPA config file, if not provided by the caller,
+ *  and creates the intial connection and manages the connection wait, close mechanisms.
+ */
+void createSocketConnection(void *config_in, void (* initKeypress)());
+   
 int createNopollConnection(noPollCtx *);
+
+/**
+ * @brief Interface to terminate WebSocket client connections and clean up resources.
+ */
 void close_and_unref_connection(noPollConn *);
-int sendResponse(noPollConn * conn,void *str, size_t bufferSize);
-void setMessageHandlers();
 
 noPollConn *get_global_conn(void);
 void set_global_conn(noPollConn *);
-
-extern pthread_mutex_t g_mutex;
-extern pthread_cond_t g_cond;
-extern pthread_mutex_t close_mut;
-extern bool close_retry;
-extern bool LastReasonStatus;
-extern char *reconnect_reason;
-
-extern  volatile unsigned int heartBeatTimer;
-extern  ParodusMsg *ParodusMsgQ;
-extern UpStreamMsg *UpStreamMsgQ;
-
-extern int numOfClients;
-
 
 #ifdef __cplusplus
 }
