+++ src/service_alive.c
@@ -1,25 +1,24 @@
 /**
- * @file internal.c
+ * @file service_alive.c
  *
- * @description This file is used to manage internal functions of parodus
+ * @description This file is used to manage keep alive section
  *
  * Copyright (c) 2015  Comcast
  */
 
 #include "ParodusInternal.h"
-#include "wss_mgr.h"
 #include "connection.h"
 #include "client_list.h"
 #include "service_alive.h"
-#include <nanomsg/nn.h>
-#include <nanomsg/pipeline.h>
 
+/*----------------------------------------------------------------------------*/
+/*                                   Macros                                   */
+/*----------------------------------------------------------------------------*/
 #define KEEPALIVE_INTERVAL_SEC                         	30
 
 /*----------------------------------------------------------------------------*/
-/*                             Internal functions                             */
+/*                             External Functions                             */
 /*----------------------------------------------------------------------------*/
-
 /*
  * @brief To handle registered services to indicate that the service is still alive.
  */
