+++ src/service_alive.h
@@ -1,14 +1,13 @@
 /**
- * @file internal.h
+ * @file service_alive.h
  *
- * @description This file is used to manage internal functions of parodus
+ * @description This file is used to manage keep alive section
  *
  * Copyright (c) 2015  Comcast
  */
+ 
 #ifndef _SERVICE_ALIVE_H_
 #define _SERVICE_ALIVE_H_
-
-#include <stdio.h>
 
 /*----------------------------------------------------------------------------*/
 /*                             Function Prototypes                            */
