+++ src/string_helpers.c
@@ -14,7 +14,7 @@
  * limitations under the License.
  *
  */
-#include "ParodusInternal.h"
+#include "string.h"
 
 /*----------------------------------------------------------------------------*/
 /*                                   Macros                                   */
