--- tests/nopoll_mock.h
@@ -1,16 +1,0 @@
-bool LastReasonStatus;
-bool close_retry;
-char *reconnect_reason;
-volatile unsigned int heartBeatTimer;
-void *metadataPack;
-size_t metaPackSize;
-pthread_mutex_t close_mut;
-pthread_mutex_t g_mutex;
-pthread_cond_t g_cond;
-
-ParodusCfg gPC;
-noPollConn gNPConn;
-
-ParodusMsg gMsg;
-ParodusMsg *ParodusMsgQ = &gMsg;
-
